// MedicalJournal-frontend/src/app/models/MeasurementReminderCourse.ts
export class MeasurementReminderCourse{
    idMeasurementType: number;
    idMeasurementReminder: string;
    //idMeasurementValueType: number;
    startDate: Date;
    idCycle: string;
    idHavingMealsType: Number;
    havingMealsTime: Number;
    annotation: string;
    isActive: number;
    reminderTimes: Date[];
}